Added Controllers for the routes and converted some syntax to ES6+
